+++ crates/oxide_emit/src/expr/object/mod.rs
@@ -7,7 +7,7 @@ mod array;
 mod object_literal;
 
 impl Emitter {
-    pub(crate) fn emit_object_domain(&self, expr: &Expression, ctx: &mut CompileCtx) -> Result<u8, String> {
+    pub(crate) fn emit_object_domain(&self, expr: &Expression, ctx: &mut CompileCtx) -> Result<u32, String> {
         match expr {
             Expression::ObjectExpression(obj) => self.emit_object_expression(obj, ctx),
             Expression::ArrayExpression(arr) => self.emit_array_expression(arr, ctx),
